More stuff. Just had to roll back on a huge mistake, thought it would be a safe thing to make a commit now that everything's back in order

+++ TowerAttack/TAEnemy.h
@@ -14,8 +14,8 @@
 @interface TAEnemy : TAUnit
 
 @property (nonatomic) CGFloat movementSpeed;
-@property (nonatomic) NSInteger maximumHealth;
-@property (nonatomic) NSInteger currentHealth;
+@property (nonatomic) CGFloat maximumHealth;
+@property (nonatomic) CGFloat currentHealth;
 @property (nonatomic, strong) SKSpriteNode *healthBarInside;
 @property (nonatomic) NSUInteger goldReward;
